refactor(mcp-server): TS port of HTTP server + auth middleware (T4.1)
Implements T4.1 of specs/maintainability-overhaul-tasks.md — Phase 4
kickoff. The HTTP surface ports from a single 434-LOC JS bin to TS
modules with a clean composition root.

New TS modules under packages/mcp-server/src/http/:
- auth.ts — `authenticateMcp`, `isAllowedOrigin`, `parseAgentTokenMap`,
  `parseCsv`, `timingSafeEqual`. Pure functions over an `AuthConfig`
  record; `parseAgentTokenMap` throws `AgentTokensError` instead of
  calling `process.exit` so it stays unit-testable.
- routes.ts — `createRouteHandler(deps)` returning the request
  dispatcher. Closes over the store + auth + publicDir + body limit.
  Inlines the HTTP IO helpers (`sendJson`, `sendFile`, `sendEmpty`,
  `sendUnauthorized`, `readJson`).
- server.ts — `createHttpServer(options): http.Server`. The composition
  root: wires auth + routes onto `http.createServer`.

New TS entrypoint at src/bin/http.ts. Owns env parsing + boot-time
validation + signal handling; calls `createHttpServer` and listens.

Tests: 32 HTTP tests ported from node:test (http.test.js +
sessions.http.test.js) to Vitest under tests/http/. Behaviour coverage
is identical — the tests still spawn the compiled bin and exercise the
HTTP surface end-to-end.

Package wiring:
- package.json: added `build`/`prepare`/`test:node`/`test:vitest`
  scripts; `serve` now runs `dist/bin/http.js`. Added `@types/node`.
- vitest.config.ts: include `tests/**/*.test.ts`, serial fork pool,
  30s timeout (HTTP tests spawn child processes), and the same
  `node:sqlite` externalize pattern @librarian/core uses.

External callers updated to the new dist path: `test/helpers.js`,
`scripts/healthcheck.js`, `scripts/smoke-test.js`, `Dockerfile` (now
installs full deps → builds → prunes prod).

Stop-gap: `src/mcp/dispatch.d.ts` ambient declarations cover the
remaining JS dispatch module so routes.ts can import it with full
types. Deleted in T4.2 when dispatch ports to TS.

Verification:
- typecheck clean across core / mcp-server / cli / dashboard
- 80 node:test + 116 vitest = 196 tests (floor 177)
- lint + Prettier clean
- check:test-count, check:storage-fixture, check:schema-version all OK
- healthcheck 5/5, smoke pass, claude-code wrapper smoke OK
