+++ packages/mcp-server/package.json
@@ -10,15 +10,20 @@
     "./dispatch": "./src/mcp/dispatch.js"
   },
   "scripts": {
+    "prepare": "tsc -p tsconfig.json",
+    "build": "tsc -p tsconfig.json",
     "start": "node --no-warnings src/bin/stdio.js",
-    "serve": "node --no-warnings src/bin/http.js",
-    "test": "node --no-warnings --test tests/*.test.js",
+    "serve": "node --no-warnings dist/bin/http.js",
+    "test:node": "node --no-warnings --test tests/*.test.js",
+    "test:vitest": "vitest run",
+    "test": "pnpm run build && pnpm run test:node && pnpm run test:vitest",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "@librarian/core": "workspace:*"
   },
   "devDependencies": {
+    "@types/node": "^22.10.5",
     "typescript": "catalog:",
     "vitest": "catalog:"
   },
